requirements.txt
# Core Python packages
pandas==2.1.4
numpy==1.24.3
python-dotenv==1.0.0

# Azure SDK packages
azure-storage-blob==12.17.0
azure-keyvault-secrets==4.7.0
azure-identity==1.14.0
azure-mgmt-databricks==1.0.0

# Spark and Delta Lake
pyspark==3.4.1
delta-spark==2.4.0

# dbt packages
dbt-core==1.6.0
dbt-databricks==1.6.0

# Data quality and validation
great-expectations==0.17.23
pydantic==1.10.12

# Utilities
requests==2.31.0
click==8.1.7
PyYAML==6.0.1
Jinja2==3.1.2

# Logging and monitoring
structlog==23.1.0
colorama==0.4.6

# Development and testing
pytest==7.4.0
pytest-cov==4.1.0
black==23.7.0
flake8==6.0.0
mypy==1.5.1

# Jupyter for development
jupyter==1.0.0
notebook==7.0.2

# Additional utilities
tqdm==4.65.0
python-dateutil==2.8.2
